basics of cloudwatch client api
