+++ auto-matching-api-rails/.env.sample
@@ -2,6 +2,10 @@ ROLLBAR_ACCESS_TOKEN = ""
 NEW_RELIC_LICENCE_KEY = ""
 SLACK_WEBHOOK_URL = ""
 
+# AWS
+AWS_ACCESS_KEY_ID = ""
+AWS_SECRET_ACCESS_KEY = ""
+
 # Redis
 REDIS_HOST = redis
 REDIS_PORT = 6379
