+++ auto-matching-api-rails/app/jobs/post_slack_job.rb
@@ -2,7 +2,7 @@ class PostSlackJob < AutoJob
   def perform(sender_class_s, area_list = nil, genre_list = nil)
     klass = sender_class_s.constantize
     key = klass.new.source_site_key
-    util = AutoMatching::SlackUtils.new(time_now)
+    util = AutoMatching::Client::SlackUtils.new(time_now)
 
     AutoMatching::Reader::Executor.new.run(klass, area_list, genre_list)
